fix: use db.count instead of get_all
